requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
